+++ apps/extension/src/pages/Home/index.tsx
@@ -1,14 +1,25 @@
 import type { FC } from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router";
-import { Button, Heading, HStack, IconButton } from "@chakra-ui/react";
+import { useTranslation } from "react-i18next";
+import {
+  Box,
+  Button,
+  Heading,
+  HStack,
+  Stack,
+  IconButton,
+  Text,
+} from "@chakra-ui/react";
+import { FaCircleInfo } from "react-icons/fa6";
 import { RiSettings5Fill } from "react-icons/ri";
 import { Routes } from "@constants";
 import { Header } from "@components";
-import { changeTheme } from "@reducers/settingsReducer";
-import { FaCircleInfo } from "react-icons/fa6";
+import { LastIP } from "@components/LastIP";
 
 export const HomePage: FC = () => {
+  const { t } = useTranslation();
+
   const navigate = useNavigate();
   const dispatch = useDispatch();
 
@@ -23,7 +34,7 @@ export const HomePage: FC = () => {
   return (
     <>
       <Header>
-        <Heading>Romania, Cluj-Napoca</Heading>
+        <Heading size="lg">Romania, Cluj-Napoca</Heading>
         <HStack>
           <IconButton
             size="md"
@@ -43,13 +54,42 @@ export const HomePage: FC = () => {
           </IconButton>
         </HStack>
       </Header>
-      <br />
-      <Button onClick={() => dispatch(changeTheme({ theme: "dark" }))}>
-        Change to dark
-      </Button>
-      <Button onClick={() => dispatch(changeTheme({ theme: "light" }))}>
-        Change to light
-      </Button>
+      <Box display="flex" flexDirection="column" gap="6" paddingY="5">
+        <Box
+          id="general"
+          display="flex"
+          flexDirection="column"
+          gap="3"
+          fontSize="lg"
+        >
+          <Stack>
+            <Text fontWeight="bold">{`${t("yourPublicIP")}: `}</Text>
+            <Button width="100%">
+              2a02:2f0e:3d12:7c00:1156:9423:cbaf:7cc7
+            </Button>
+          </Stack>
+          <Stack>
+            <Text fontWeight="bold">{`${t("yourPrivateIP")}: `}</Text>
+            <Button width="100%">192.168.1.2</Button>
+          </Stack>
+        </Box>
+        <Box
+          id="general"
+          display="flex"
+          flexDirection="column"
+          gap="3"
+          fontSize="lg"
+        >
+          <Stack>
+            <Text fontWeight="bold">{`${t("lastIPs")}: `}</Text>
+            <LastIP />
+            <LastIP />
+            <LastIP />
+            <LastIP />
+            <LastIP />
+          </Stack>
+        </Box>
+      </Box>
     </>
   );
 };